qemu_chr_be_write uses size_t
all use size_t arg except ... where we cast

    if (len <= room) {
        /* Fits */
        qemu_chr_be_write(chr, io_buf, len);
chardev/baum.c:304:        qemu_chr_be_write(chr, io_buf, len);

    if (size > 0) {
chardev/char-win.c:55:        qemu_chr_be_write(chr, buf, size);
chardev/char-fd.c:66:        qemu_chr_be_write(chr, buf, ret);

    if (ret <= 0) {
        pty_chr_state(chr, 0);
        return FALSE;
    } else {
        pty_chr_state(chr, 1);
        qemu_chr_be_write(chr, buf, ret);
chardev/char-pty.c:181:        qemu_chr_be_write(chr, buf, ret);

    while (len > 0) {
        size_t can_write = qemu_chr_be_can_write(chr);
        last_out = MIN(len, can_write);
chardev/spice.c:51:        qemu_chr_be_write(chr, p, last_out);

+++ chardev/baum.c
@@ -301,7 +301,7 @@ static void baum_write_packet(BaumChardev *baum, const uint8_t *buf, size_t len)
     len = cur - io_buf;
     if (len <= room) {
         /* Fits */
-        qemu_chr_be_write(chr, io_buf, len);
+        qemu_chr_be_write(chr, io_buf, (size_t)len);
     } else {
         int first;
         uint8_t out;
